Move physics value grid interface to its own file
